加载博客细节接口增加了作者头像的返回 修改了取消点踩博客接口的bug 增加了查看是否收藏过博客的接口 收藏博客时可以正确的增加博客的收藏量 增加了删除消息的接口 /message/list接口现在会把返回的信息列为已读 修改了关注与未关注查询相反的bug 在删除收藏夹前会删除收藏夹的收藏博客关系 增加了一些权限检查

+++ Lunar-framework/src/main/java/com/lunar/service/impl/FolderServiceImpl.java
@@ -77,7 +77,13 @@ public class FolderServiceImpl extends ServiceImpl<FolderMapper, Folder> impleme
 
     @Override
     public ResponseResult deleteFolder(Integer folderId) {
-        //TODO: 检查权限
+        //获取token中的userId
+        Integer userId = UserFillUtils.getUserIdFromToken();
+
+        //如果没有找到userId 返回需要登陆
+        if(Objects.isNull(userId)) {
+            return ResponseResult.errorResult(AppHttpCodeEnum.NEED_LOGIN.getCode(), AppHttpCodeEnum.NEED_LOGIN.getMsg());
+        }
         LambdaQueryWrapper<FolderCollect> folderCollectLambdaQueryWrapper = new LambdaQueryWrapper<>();
         folderCollectLambdaQueryWrapper.eq(FolderCollect::getFolderId, folderId);
         folderCollectService.remove(folderCollectLambdaQueryWrapper);
@@ -89,7 +95,13 @@ public class FolderServiceImpl extends ServiceImpl<FolderMapper, Folder> impleme
 
     @Override
     public ResponseResult updateFolderName(Integer folderId, String folderName) {
-        //TODO: 检查权限
+        //获取token中的userId
+        Integer userId = UserFillUtils.getUserIdFromToken();
+
+        //如果没有找到userId 返回需要登陆
+        if(Objects.isNull(userId)) {
+            return ResponseResult.errorResult(AppHttpCodeEnum.NEED_LOGIN.getCode(), AppHttpCodeEnum.NEED_LOGIN.getMsg());
+        }
 
         //根据id查找收藏夹
         Folder folder = getById(folderId);
